# docs/swagger.yaml
components:
  schemas:
...
    analysis.Coverage:
      properties:
        reasons:
          items:
            type: string
          type: array
          uniqueItems: false
        semantic_complete:
          type: boolean
        syntax_complete:
          type: boolean
      type: object
    analysis.Dependencies:
      properties:
        data_models:
          items:
            type: string
          type: array
          uniqueItems: false
        datasets:
          items:
            type: string
          type: array
          uniqueItems: false
        indexes:
          items:
            type: string
          type: array
          uniqueItems: false
        lookups:
          items:
            type: string
          type: array
          uniqueItems: false
        macros:
          items:
            type: string
          type: array
          uniqueItems: false
        source_types:
          items:
            type: string
          type: array
          uniqueItems: false
        sources:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    analysis.Diagnostic:
      properties:
        category:
          type: string
        code:
          type: string
        location:
          $ref: '#/components/schemas/analysis.Location'
        message:
          type: string
        scope_id:
          type: string
        severity:
          type: string
        stage_id:
          type: string
      type: object
    analysis.FieldBinding:
      properties:
        conditional:
          type: boolean
        name:
          type: string
        origin_reference_ids:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    analysis.FieldState:
      properties:
        fields:
          items:
            $ref: '#/components/schemas/analysis.FieldBinding'
          type: array
          uniqueItems: false
        open:
          type: boolean
        removed:
          items:
            type: string
          type: array
          uniqueItems: false
        uncertain:
          type: boolean
      type: object
    analysis.Lineage:
      properties:
        after:
          $ref: '#/components/schemas/analysis.FieldState'
        before:
          $ref: '#/components/schemas/analysis.FieldState'
        execution_order:
          type: integer
        phase:
          type: string
        scope_id:
          type: string
        stage_id:
          type: string
        transitions:
          items:
            $ref: '#/components/schemas/analysis.Transition'
          type: array
          uniqueItems: false
      type: object
    analysis.Location:
      properties:
        end:
          $ref: '#/components/schemas/analysis.Position'
        start:
          $ref: '#/components/schemas/analysis.Position'
      type: object
    analysis.Position:
      properties:
        column:
          type: integer
        line:
          type: integer
        offset:
          type: integer
      type: object
    analysis.QueryDocument:
      properties:
        language:
          type: string
        profile:
          type: string
        source_id:
          type: string
        text:
          type: string
        version:
          type: string
      type: object
    analysis.Reference:
      properties:
        binding:
          type: string
        id:
          type: string
        kind:
          type: string
        location:
          $ref: '#/components/schemas/analysis.Location'
        normalized_name:
          type: string
        origin_reference_ids:
          items:
            type: string
          type: array
          uniqueItems: false
        original_name:
          type: string
        resolution:
          type: string
        role:
          type: string
        scope_id:
          type: string
        stage_id:
          type: string
      type: object
    analysis.Result:
      properties:
        coverage:
          $ref: '#/components/schemas/analysis.Coverage'
        dependencies:
          $ref: '#/components/schemas/analysis.Dependencies'
        diagnostics:
          items:
            $ref: '#/components/schemas/analysis.Diagnostic'
          type: array
          uniqueItems: false
        document:
          $ref: '#/components/schemas/analysis.QueryDocument'
        lineage:
          items:
            $ref: '#/components/schemas/analysis.Lineage'
          type: array
          uniqueItems: false
        references:
          items:
            $ref: '#/components/schemas/analysis.Reference'
          type: array
          uniqueItems: false
        schema_version:
          type: integer
        scopes:
          items:
            $ref: '#/components/schemas/analysis.Scope'
          type: array
          uniqueItems: false
        stages:
          items:
            $ref: '#/components/schemas/analysis.Stage'
          type: array
          uniqueItems: false
        status:
          $ref: '#/components/schemas/analysis.Status'
      type: object
...
    analysis.Scope:
      properties:
        id:
          type: string
        kind:
          type: string
        location:
          $ref: '#/components/schemas/analysis.Location'
        parent_id:
          type: string
        stage_id:
          type: string
      type: object
    analysis.Stage:
      properties:
        command:
          type: string
        id:
          type: string
        location:
          $ref: '#/components/schemas/analysis.Location'
        position:
          type: integer
        scope_id:
          type: string
        semantic_complete:
          type: boolean
      type: object
    analysis.Status:
      type: string
      x-enum-varnames:
      - Valid
      - Invalid
      - Incomplete
    analysis.Transition:
      properties:
        conditional:
          type: boolean
        input_reference_ids:
          items:
            type: string
          type: array
          uniqueItems: false
        operation:
          type: string
        output:
          type: string
        output_reference_id:
          type: string
      type: object
...
    api.ErrorResponse:
      description: Standard error response
      properties:
        code:
          description: HTTP status code
          example: 400
          type: integer
          x-order: '3'
        error:
          description: Always true for error responses
          example: true
          type: boolean
          x-order: '1'
        message:
          description: Human-readable error message
          example: Invalid request
          type: string
          x-order: '2'
      type: object
...
    api.RewriteBatchRequest:
      additionalProperties: false
      properties:
        documents:
          items:
            $ref: '#/components/schemas/validation.QueryDocument'
          minItems: 1
          type: array
        mode:
          default: preview
          enum:
          - preview
          - apply
          type: string
        rules:
          items:
            $ref: '#/components/schemas/api.RewriteRule'
          type: array
        schema_version:
          const: 1
          type: integer
        validation_target:
          description: Optional inline offline validation target. Paths and retrieval
            URLs are not accepted.
          oneOf:
          - additionalProperties: false
            properties:
              catalog:
                oneOf:
                - items:
                    minLength: 1
                    type: string
                  type: array
                  uniqueItems: true
                - $ref: '#/components/schemas/validation.FieldCatalog'
              kind:
                const: field_list
            required:
            - kind
            - catalog
            type: object
          - additionalProperties: false
            properties:
              base_uri:
                minLength: 1
                pattern: \S
                type: string
              identity:
                minLength: 1
                pattern: \S
                type: string
              kind:
                const: json_schema
              resources:
                additionalProperties:
                  oneOf:
                  - type: object
                  - type: boolean
                description: URI-keyed local inline JSON Schema resources. No URI
                  is fetched.
                propertyNames:
                  minLength: 1
                  pattern: \S
                  type: string
                type: object
              schema:
                oneOf:
                - type: object
                - type: boolean
            required:
            - kind
            - schema
            type: object
          - additionalProperties: false
            properties:
              catalog:
                description: Official compiled OCSF catalog with compile_version 1,
                  exact version and full extension set. Catalog vocabulary and annotation
                  members are allowed.
                type: object
              identity:
                minLength: 1
                pattern: \S
                type: string
              kind:
                const: ocsf
              selection:
                $ref: '#/components/schemas/api.SchemaValidationSelection'
            required:
            - kind
            - catalog
            - selection
            type: object
      required:
      - schema_version
      - documents
      - rules
      type: object
    api.RewriteCondition:
      additionalProperties: false
      oneOf:
      - required:
        - all
      - required:
        - any
      - oneOf:
        - properties:
            fact:
              const: literal
          required:
          - operator
          - value
        - not:
            anyOf:
            - required:
              - operator
            - required:
              - value
          properties:
            fact:
              const: source_reference_present
        required:
        - fact
        - kind
        - identity
      properties:
        all:
          items: &id001
            $ref: '#/components/schemas/api.RewriteCondition'
          minItems: 1
          type: array
        any:
          items: *id001
          minItems: 1
          type: array
        fact:
          enum:
          - literal
          - source_reference_present
          type: string
        identity:
          $ref: '#/components/schemas/api.RewriteIdentity'
        kind:
          enum:
          - field
          - index
          - source
          - sourcetype
          - lookup
          - dataset
          - data_model
          type: string
        operator:
          enum:
          - equals
          - contains
          type: string
        value:
          oneOf:
          - type: string
          - type: number
          - type: boolean
          - type: 'null'
      type: object
    api.RewriteIdentity:
      additionalProperties: false
      oneOf:
      - required:
        - name
      - required:
        - path
      properties:
        name:
          minLength: 1
          pattern: \S
          type: string
        path:
          items:
            minLength: 1
            pattern: \S
            type: string
          minItems: 1
          type: array
      type: object
    api.RewriteRequest:
      additionalProperties: false
      properties:
        document:
          $ref: '#/components/schemas/validation.QueryDocument'
        mode:
          default: preview
          enum:
          - preview
          - apply
          type: string
        rules:
          items:
            $ref: '#/components/schemas/api.RewriteRule'
          type: array
        schema_version:
          const: 1
          type: integer
        validation_target:
          description: Optional inline offline validation target. Paths and retrieval
            URLs are not accepted.
          oneOf:
          - additionalProperties: false
            properties:
              catalog:
                oneOf:
                - items:
                    minLength: 1
                    type: string
                  type: array
                  uniqueItems: true
                - $ref: '#/components/schemas/validation.FieldCatalog'
              kind:
                const: field_list
            required:
            - kind
            - catalog
            type: object
          - additionalProperties: false
            properties:
              base_uri:
                minLength: 1
                pattern: \S
                type: string
              identity:
                minLength: 1
                pattern: \S
                type: string
              kind:
                const: json_schema
              resources:
                additionalProperties:
                  oneOf:
                  - type: object
                  - type: boolean
                description: URI-keyed local inline JSON Schema resources. No URI
                  is fetched.
                propertyNames:
                  minLength: 1
                  pattern: \S
                  type: string
                type: object
              schema:
                oneOf:
                - type: object
                - type: boolean
            required:
            - kind
            - schema
            type: object
          - additionalProperties: false
            properties:
              catalog:
                description: Official compiled OCSF catalog with compile_version 1,
                  exact version and full extension set. Catalog vocabulary and annotation
                  members are allowed.
                type: object
              identity:
                minLength: 1
                pattern: \S
                type: string
              kind:
                const: ocsf
              selection:
                $ref: '#/components/schemas/api.SchemaValidationSelection'
            required:
            - kind
            - catalog
            - selection
            type: object
      required:
      - schema_version
      - document
      - rules
      type: object
    api.RewriteRule:
      additionalProperties: false
      properties:
        id:
          minLength: 1
          pattern: \S
          type: string
        kind:
          enum:
          - field
          - index
          - source
          - sourcetype
          - lookup
          - dataset
          - data_model
          type: string
        source:
          $ref: '#/components/schemas/api.RewriteIdentity'
        target:
          $ref: '#/components/schemas/api.RewriteIdentity'
        when:
          $ref: '#/components/schemas/api.RewriteCondition'
      required:
      - id
      - kind
      - source
      - target
      type: object
...
    api.SchemaValidationSelection:
      additionalProperties: false
      description: Exact version and exactly one selector. Omitted profiles/extensions
        normalize to empty arrays; null is invalid. Names are nonblank UTF-8, unique
        and case-sensitive.
      oneOf:
      - required:
        - class
      - required:
        - class_uid
      - required:
        - category
      - required:
        - category_uid
      properties:
        category:
          minLength: 1
          pattern: \S
          type: string
        category_uid:
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          type: integer
        class:
          minLength: 1
          pattern: \S
          type: string
        class_uid:
          maximum: 9223372036854775807
          minimum: -9223372036854775808
          type: integer
        extensions:
          items:
            minLength: 1
            pattern: \S
            type: string
          type: array
          uniqueItems: true
        profiles:
          items:
            minLength: 1
            pattern: \S
            type: string
          type: array
          uniqueItems: true
        version:
          minLength: 1
          pattern: \S
          type: string
      required:
      - version
      type: object
...
    rewrite.BatchResult:
      properties:
        reports:
          items:
            $ref: '#/components/schemas/rewrite.Result'
          type: array
          uniqueItems: false
        schema_version:
          type: integer
        status:
          type: string
          x-enum-varnames:
          - Valid
          - Invalid
          - Incomplete
      type: object
    rewrite.CandidateValidation:
      properties:
        field_list:
          $ref: '#/components/schemas/validation.Report'
        kind:
          type: string
        schema:
          $ref: '#/components/schemas/validation.SchemaReport'
      type: object
    rewrite.Change:
      properties:
        candidate_applied:
          type: boolean
        candidate_location:
          $ref: '#/components/schemas/analysis.Location'
        candidate_reference_ids:
          items:
            type: string
          type: array
          uniqueItems: false
        committed:
          type: boolean
        group_id:
          type: string
        new_text:
          type: string
        old_text:
          type: string
        original_location:
          $ref: '#/components/schemas/analysis.Location'
        original_reference_ids:
          items:
            type: string
          type: array
          uniqueItems: false
        outcome:
          type: string
        reason:
          type: string
        rule_ids:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    rewrite.ConditionEvaluation:
      properties:
        children:
          items:
            $ref: '#/components/schemas/rewrite.ConditionEvaluation'
          type: array
          uniqueItems: false
        reason:
          type: string
        reference_ids:
          items:
            type: string
          type: array
          uniqueItems: false
        state:
          type: string
      type: object
    rewrite.Coverage:
      properties:
        reasons:
          items:
            type: string
          type: array
          uniqueItems: false
        rewrite_complete:
          type: boolean
        semantic_complete:
          type: boolean
        syntax_complete:
          type: boolean
        validation:
          $ref: '#/components/schemas/validation.Coverage'
      type: object
...
    rewrite.Result:
      properties:
        candidate_analysis:
          $ref: '#/components/schemas/analysis.Result'
        candidate_text:
          type: string
        candidate_validation:
          $ref: '#/components/schemas/rewrite.CandidateValidation'
        changes:
          items:
            $ref: '#/components/schemas/rewrite.Change'
          type: array
          uniqueItems: false
        committed:
          type: boolean
        coverage:
          $ref: '#/components/schemas/rewrite.Coverage'
        document:
          $ref: '#/components/schemas/analysis.QueryDocument'
        mode:
          type: string
          x-enum-varnames:
          - Preview
          - Apply
        original_analysis:
          $ref: '#/components/schemas/analysis.Result'
        original_text:
          type: string
        rule_evaluations:
          items:
            $ref: '#/components/schemas/rewrite.RuleEvaluation'
          type: array
          uniqueItems: false
        schema_version:
          type: integer
        status:
          type: string
          x-enum-varnames:
          - Valid
          - Invalid
          - Incomplete
        text:
          type: string
      type: object
    rewrite.RuleEvaluation:
      properties:
        condition:
          $ref: '#/components/schemas/rewrite.ConditionEvaluation'
        location:
          $ref: '#/components/schemas/analysis.Location'
        outcome:
          type: string
        reason:
          type: string
        reference_ids:
          items:
            type: string
          type: array
          uniqueItems: false
        rule_id:
          type: string
      type: object
...
    validation.Coverage:
      properties:
        reasons:
          items:
            type: string
          type: array
          uniqueItems: false
        schema_complete:
          type: boolean
        semantic_complete:
          type: boolean
        syntax_complete:
          type: boolean
      type: object
    validation.FieldCatalog:
      additionalProperties: false
      properties:
        fields:
          items:
            minLength: 1
            type: string
          type: array
          uniqueItems: true
        identity:
          type: string
        optional_fields:
          items:
            minLength: 1
            type: string
          type: array
          uniqueItems: true
        version:
          type: string
      required:
      - fields
      type: object
    validation.Match:
      properties:
        binding:
          type: string
        name:
          type: string
        outcome:
          type: string
      type: object
...
    validation.QueryDocument:
      additionalProperties: false
      properties:
        language:
          default: spl
          description: Omitted or empty selects the default; other values are input
            errors.
          enum:
          - ''
          - spl
          - spl2
          type: string
        profile:
          default: splunkd
          description: Omitted or empty selects the default; other values are input
            errors.
          enum:
          - ''
          - splunkd
          type: string
        source_id:
          type: string
        text:
          type: string
        version:
          default: current
          description: Omitted or empty selects the default; other values are input
            errors.
          enum:
          - ''
          - current
          type: string
      required:
      - text
      type: object
    validation.ReferenceOutcome:
      properties:
        matches:
          items:
            $ref: '#/components/schemas/validation.Match'
          type: array
          uniqueItems: false
        outcome:
          type: string
        reference_id:
          type: string
      type: object
    validation.Report:
      properties:
        analysis:
          $ref: '#/components/schemas/analysis.Result'
        coverage:
          $ref: '#/components/schemas/validation.Coverage'
        diagnostics:
          items:
            $ref: '#/components/schemas/analysis.Diagnostic'
          type: array
          uniqueItems: false
        outcomes:
          items:
            $ref: '#/components/schemas/validation.ReferenceOutcome'
          type: array
          uniqueItems: false
        schema_version:
          type: integer
        status:
          type: string
          x-enum-varnames:
          - Valid
          - Invalid
          - Incomplete
        target:
          $ref: '#/components/schemas/validation.Target'
      type: object
...
    validation.SchemaClass:
      properties:
        key:
          type: string
        uid:
          type: integer
      type: object
    validation.SchemaEvidence:
      properties:
        branch:
          type: string
        class_key:
          type: string
        class_uid:
          type: integer
        declaration_basis:
          type: string
        keyword:
          type: string
        operator:
          type: string
        pointer:
          type: string
        reason:
          type: string
        requirement:
          type: string
        resource_uri:
          type: string
      type: object
    validation.SchemaExtension:
      properties:
        key:
          type: string
        uid:
          type: integer
        version:
          type: string
      type: object
    validation.SchemaMatch:
      properties:
        binding:
          type: string
        evidence:
          items:
            $ref: '#/components/schemas/validation.SchemaEvidence'
          type: array
          uniqueItems: false
        name:
          type: string
        outcome:
          type: string
      type: object
    validation.SchemaReferenceOutcome:
      properties:
        evidence:
          items:
            $ref: '#/components/schemas/validation.SchemaEvidence'
          type: array
          uniqueItems: false
        indeterminate_classes:
          items:
            $ref: '#/components/schemas/validation.SchemaClass'
          type: array
          uniqueItems: false
        matches:
          items:
            $ref: '#/components/schemas/validation.SchemaMatch'
          type: array
          uniqueItems: false
        matches_complete:
          type: boolean
        missing_classes:
          items:
            $ref: '#/components/schemas/validation.SchemaClass'
          type: array
          uniqueItems: false
        outcome:
          type: string
        reference_id:
          type: string
        supporting_classes:
          items:
            $ref: '#/components/schemas/validation.SchemaClass'
          type: array
          uniqueItems: false
      type: object
    validation.SchemaReport:
      properties:
        analysis:
          $ref: '#/components/schemas/analysis.Result'
        coverage:
          $ref: '#/components/schemas/validation.Coverage'
        diagnostics:
          items:
            $ref: '#/components/schemas/analysis.Diagnostic'
          type: array
          uniqueItems: false
        outcomes:
          items:
            $ref: '#/components/schemas/validation.SchemaReferenceOutcome'
          type: array
          uniqueItems: false
        schema_version:
          type: integer
        status:
          type: string
          x-enum-varnames:
          - Valid
          - Invalid
          - Incomplete
        target:
          $ref: '#/components/schemas/validation.SchemaTargetInfo'
      type: object
    validation.SchemaTargetInfo:
      properties:
        base_uri:
          type: string
        compile_version:
          type: integer
        dialect:
          type: string
        extensions:
          items:
            $ref: '#/components/schemas/validation.SchemaExtension'
          type: array
          uniqueItems: false
        identity:
          type: string
        kind:
          type: string
        limitations:
          items:
            type: string
          type: array
          uniqueItems: false
        members:
          items:
            $ref: '#/components/schemas/validation.SchemaClass'
          type: array
          uniqueItems: false
        resource_uris:
          items:
            type: string
          type: array
          uniqueItems: false
        selection:
          $ref: '#/components/schemas/validation.OCSFSelection'
        version:
          type: string
      type: object
    validation.Target:
      properties:
        fields:
          items:
            type: string
          type: array
          uniqueItems: false
        identity:
          type: string
        kind:
          type: string
        optional_fields:
          items:
            type: string
          type: array
          uniqueItems: false
        version:
          type: string
      type: object
...
info:
  contact:
    email: support@example.com
    name: SPL Toolkit Support
  description: Offline field mapping, flat discovery, and validation for the supported
    Milestone 1 SPL grammar.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: SPL Toolkit API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /query/rewrite:
    post:
      description: Strict versioned request with one query document, explicit inline
        rules, optional inline validation target, and preview mode by default. No
        filesystem paths, URL retrieval, or mapper configuration. Query statuses valid,
        invalid, and incomplete all return HTTP 200. Body limit is 8 MiB.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.RewriteRequest'
        description: Versioned single-query rewrite request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rewrite.Result'
          description: Canonical rewrite report
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Input error, content type, or body limit
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unexpected internal error
      summary: Preview or apply explicit safe rewrite rules
      tags:
      - query
  /query/rewrite/batch:
    post:
      description: Strict versioned request with nonempty query documents, explicit
        inline rules, optional inline validation target, and preview mode by default.
        No filesystem paths, URL retrieval, or mapper configuration. Query statuses
        valid, invalid, and incomplete all return HTTP 200. Body limit is 8 MiB.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/api.RewriteBatchRequest'
        description: Versioned ordered-batch rewrite request
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rewrite.BatchResult'
          description: Canonical ordered rewrite reports
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Input error, content type, or body limit
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorResponse'
          description: Unexpected internal error
      summary: Preview or apply explicit safe rewrite rules to an ordered batch
      tags:
      - query
...
servers:
- description: Development server
  url: http://localhost:8080/api/v1
...